Fixed envs import bug
