+++ envs/spring.py
@@ -1,37 +1,35 @@
 import numpy as np
 
-from .environments import Environment, visualize_rollout
+from environments import Environment, visualize_rollout
 
 
-class Pendulum(Environment):
-    """Pendulum System
+class Spring(Environment):
+    """Spring System
 
     Equations of movement are:
 
-        theta'' = -(g/l)*sin(theta)
+        x'' = -(k/m)*x
 
     """
-    def __init__(self, mass, length, g, q=None, p=None):
-        """Constructor for pendulum system
+    def __init__(self, mass, elastic_cst, q=None, p=None):
+        """Constructor for spring system
 
         Args:
-            mass (float): Pendulum mass (kg)
-            length (float): Pendulum length (m)
-            g (float): Gravity of the environment (m/s^2)
-            q ([float], optional): Generalized position in 1-D space: Phase (rad). Defaults to None
-            p ([float], optional): Generalized momentum in 1-D space: Angular momentum (kg*m^2/s). Defaults to None
+            mass (float): Spring mass
+            elastic_cst (float): Spring elastic constant
+            q ([float], optional): Generalized position in 1-D space: Position (m). Defaults to None
+            p ([float], optional): Generalized momentum in 1-D space: Linear momentum (kg*m/s). Defaults to None
         """
         self.mass = mass
-        self.length = length
-        self.g = g
+        self.elastic_cst = elastic_cst
         super().__init__(q=q, p=p)
 
     def set(self, q, p):
-        """Sets initial conditions for pendulum
+        """Sets initial conditions for spring system
 
         Args:
-            q ([float]): Generalized position in 1-D space: Phase (rad)
-            p ([float]): Generalized momentum in 1-D space: Angular momentum (kg*m^2/s)
+            q ([float]): Generalized position in 1-D space: Position (m)
+            p ([float]): Generalized momentum in 1-D space: Linear momentum (kg*m/s)
 
         Raises:
             ValueError: If p and q are not in 1-D space
@@ -54,8 +52,7 @@ class Pendulum(Environment):
         Returns:
             equations ([float]): Movement equations of the physical system
         """
-        return [(states[1] / (self.mass * self.length * self.length)),
-                -self.g * self.mass * self.length * np.sin(states[0])]
+        return [states[1] / self.mass, -self.elastic_cst * states[0]]
 
     def _draw(self, res=32, color=True, world_size=1.5):
         """Returns array of the environment evolution
@@ -78,15 +75,12 @@ class Pendulum(Environment):
         [I, J] = np.meshgrid(grid, grid)
         for t in range(length):
             if color:
-                vid[t, :, :, 0] += np.exp(-(((I - np.sin(q[t]))**2 +
-                                             (J - np.cos(q[t]))**2) /
+                vid[t, :, :, 0] += np.exp(-(((I - 0)**2 + (J - q[t])**2) /
                                             (self.mass**2))**4)
-                vid[t, :, :, 1] += np.exp(-(((I - np.sin(q[t]))**2 +
-                                             (J - np.cos(q[t]))**2) /
+                vid[t, :, :, 1] += np.exp(-(((I - 0)**2 + (J - q[t])**2) /
                                             (self.mass**2))**4)
             else:
-                vid[t, :, :, 0] += np.exp(-(((I - np.sin(q[t]))**2 +
-                                             (J - np.cos(q[t]))**2) /
+                vid[t, :, :, 0] += np.exp(-(((I - 0)**2 + (J - q[t])**2) /
                                             (self.mass**2))**4)
             vid[t][vid[t] > 1] = 1
 
@@ -106,13 +100,13 @@ class Pendulum(Environment):
 # Sample code for sampling rollouts
 if __name__ == "__main__":
 
-    pd = Pendulum(mass=.5, length=1, g=3)
-    rolls = pd.sample_random_rollouts(number_of_frames=100,
+    sp = Spring(mass=.5, elastic_cst=2)
+    rolls = sp.sample_random_rollouts(number_of_frames=100,
                                       delta_time=0.1,
                                       number_of_rollouts=16,
                                       img_size=32,
                                       noise_std=0.,
-                                      radius_bound=(1.3, 2.3),
+                                      radius_bound=(.1, 1.),
                                       world_size=1.5,
                                       seed=23)
     idx = np.random.randint(rolls.shape[0])
